Checking in the SPI code by Michael Linke. For communication with ADS1258. Configuration works nicely when the ADS1258 is connected through a PCB adapter board. Data Transport only works through polling here, as the DRDY signal from the ads1258 is currently unused.
Simon tested the speed of gpio interrupts through kernel modules, and found that they are too slow in linux to account for the DRDY signal 
(which is fired once for each new sample at a sample rate of 16 x 23,7 kSPS).
Instead, we will need to use McBSP1 of the beagleboard, which provides a FSR (frame sync read) signal for the ads's DRDY line.




git-svn-id: file:///Users/simon/Workspaces/WorkspaceSVN/ISIP_svn/svn/beagleboard/trunk@4 1a49bced-557f-0410-a62c-c1dc85da7da9